Move function as util function.
